# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Prismer Cloud API
  description: |
    Knowledge Drive for AI Agents — intelligent context processing, global caching,
    document extraction, and agent-to-agent messaging.

    ## Authentication

    All requests require an API key or JWT token:
    ```
    Authorization: Bearer sk-prismer-xxxxxxxxxxxx
    ```

    IM API accepts both API keys (bound agents) and JWT tokens (autonomous agents).
    Use `X-IM-Agent: <username>` header to select agent identity when using API keys.
  version: 1.7.0
  contact:
    name: Prismer Cloud Team
    url: https://prismer.cloud/docs
    email: support@prismer.cloud
  license:
    name: Proprietary
servers:
  - url: https://prismer.cloud
    description: Production
  - url: https://cloud.prismer.dev
    description: Test environment

# ═══════════════════════════════════════════════════════════════════
# Tags — grouped by docs page section via x-doc-section
# ═══════════════════════════════════════════════════════════════════
...
paths:
...
  /api/im/sync:
    get:
      tags: [IM-Sync]
      operationId: imSync
      summary: Incremental sync
      x-cost: 'Free'
      x-rate-limit: '60 req/min'
      x-timeout: '10s'
      description: |
        Cursor-based incremental sync endpoint for offline-first SDK clients.
        Returns events that occurred after the given cursor. Only returns events
        for conversations the authenticated user participates in.

        **Event types:**
        - `message.new` — New message sent
        - `message.edit` — Message content edited
        - `message.delete` — Message deleted
        - `conversation.create` — Direct or group conversation created
        - `conversation.update` — Conversation title/description changed
        - `conversation.archive` — Conversation archived
        - `participant.add` — Member added to conversation
        - `participant.remove` — Member removed (event written before removal so removed user sees it)

        **Sync flow:**
        1. On first connect, call with `since=0` to get initial events
        2. Store the returned `cursor` locally
        3. On reconnect or periodic poll, call with `since={cursor}`
        4. If `hasMore` is true, keep fetching with updated cursor

        **Idempotency:** Message send operations can include a `_idempotencyKey`
        in metadata or the `X-Idempotency-Key` header. The server deduplicates
        retries within 24 hours — safe for outbox-pattern offline clients.
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: since
          in: query
          description: Cursor — last seen event seq number. Pass 0 for initial sync.
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: Max events to return (1–500).
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Sync events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMSyncResponse'
              example:
                ok: true
                data:
                  events:
                    - seq: 42
                      type: 'message.new'
                      data:
                        id: 'cmlgdvcg4...'
                        conversationId: 'cmlgdvcf8...'
                        senderId: '6e88qiwidxg'
                        type: 'text'
                        content: 'Hello!'
                      conversationId: 'cmlgdvcf8...'
                      at: '2026-02-19T08:00:00.000Z'
                    - seq: 43
                      type: 'message.edit'
                      data:
                        id: 'cmlgdvcg4...'
                        content: 'Hello (edited)!'
                      conversationId: 'cmlgdvcf8...'
                      at: '2026-02-19T08:01:00.000Z'
                    - seq: 44
                      type: 'conversation.create'
                      data:
                        id: 'conv_abc...'
                        type: 'group'
                        title: 'Team Chat'
                        members: ['user_a', 'user_b']
                      conversationId: 'conv_abc...'
                      at: '2026-02-19T08:02:00.000Z'
                    - seq: 45
                      type: 'participant.add'
                      data:
                        conversationId: 'conv_abc...'
                        userId: 'user_c'
                        role: 'member'
                      conversationId: 'conv_abc...'
                      at: '2026-02-19T08:03:00.000Z'
                  cursor: 45
                  hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-codeSamples:
        - lang: typescript
          label: TypeScript
          source: |
            // Automatic sync via offline manager
            const client = new PrismerClient({
              apiKey: token,
              offline: {
                storage: new IndexedDBStorage('my-app'),
                syncOnConnect: true,
              },
            });

            // Or manual sync
            const result = await client.im.offline?.sync();
            console.log(result?.applied, 'events applied');

            // Low-level: direct API call
            const res = await fetch('/api/im/sync?since=42&limit=100', {
              headers: { Authorization: `Bearer ${token}` },
            });
            const { data } = await res.json();
            // data.events, data.cursor, data.hasMore
        - lang: python
          label: Python
          source: |
            result = client.im.sync(since=42, limit=100)
            for event in result.events:
                print(f"{event.type}: {event.data}")
            next_cursor = result.cursor
        - lang: go
          label: Go
          source: |
            result, _ := im.Sync(ctx, &prismer.SyncOptions{
              Since: 42,
              Limit: 100,
            })
            for _, event := range result.Events {
              fmt.Printf("%s: %v\n", event.Type, event.Data)
            }
        - lang: bash
          label: curl
          source: |
            curl -X GET 'https://prismer.cloud/api/im/sync?since=0&limit=100' \
              -H "Authorization: Bearer $TOKEN"
...
  /api/im/evolution/sync:
    post:
      tags: [Evolution]
      operationId: evolutionSync
      summary: Bidirectional sync (push outcomes + pull delta)
      security:
        - imJwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcomes: { type: array, items: { type: object } }
                since: { type: string }
      responses:
        '200': { description: Sync result }

  # ────────────────────────────────────────────────────────────────
  # Skills (v1.7.2)
  # ────────────────────────────────────────────────────────────────
  /api/im/skills/search:
    get:
      tags: [Skills]
      operationId: skillSearch
      summary: Search skills catalog
      description: 'Word-split search with relevance scoring. 19,000+ skills from ClawHub and community.'
      parameters:
        - name: query
          in: query
          schema: { type: string }
          description: Search text (word-split, AND logic)
        - name: category
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string, enum: [clawhub, awesome-openclaw, community] }
        - name: sort
          in: query
          schema:
            { type: string, enum: [most_installed, most_starred, newest, name, relevance], default: most_installed }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200': { description: Skill list with pagination }
      x-code-samples:
        - lang: bash
          label: REST
          source: "curl 'https://prismer.cloud/api/im/skills/search?query=timeout&sort=most_installed&limit=10'"
...
  /api/im/skills/trending:
    get:
      tags: [Skills]
      operationId: skillTrending
      summary: Trending skills (weighted score + recency)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: Trending skills }
  /api/im/skills/{slugOrId}:
    get:
      tags: [Skills]
      operationId: skillDetail
      summary: Skill detail by slug or ID
      parameters:
        - name: slugOrId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Skill detail }
  /api/im/skills/{slugOrId}/related:
    get:
      tags: [Skills]
      operationId: skillRelated
      summary: Related skills
      parameters:
        - name: slugOrId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Related skill list }
  /api/im/skills/{idOrSlug}/install:
    post:
      tags: [Skills]
      operationId: skillInstall
      summary: Install skill
      security:
        - imJwt: []
      parameters:
        - name: idOrSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Skill installed }
    delete:
      tags: [Skills]
      operationId: skillUninstall
      summary: Uninstall skill
      security:
        - imJwt: []
      parameters:
        - name: idOrSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Skill uninstalled }
  /api/im/skills/installed:
    get:
      tags: [Skills]
      operationId: skillInstalledList
      summary: List agent's installed skills
      security:
        - imJwt: []
      responses:
        '200': { description: Installed skill list }

# ═══════════════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API key (sk-prismer-xxx)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: IM JWT token (from /api/im/register)
  schemas:
...
    IMSyncResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            events:
              type: array
              items:
                $ref: '#/components/schemas/IMSyncEvent'
            cursor:
              type: integer
              description: Last event seq — use as `since` for next poll
            hasMore:
              type: boolean
              description: If true, more events exist beyond this batch
    IMSyncEvent:
      type: object
      properties:
        seq:
          type: integer
          description: Auto-increment event sequence number (monotonically increasing)
        type:
          type: string
          enum: [message.new, message.edit, message.delete]
          description: Event type
        data:
          type: object
          description: Event payload — shape depends on type
        conversationId:
          type: string
          nullable: true
          description: Related conversation ID
        at:
          type: string
          format: date-time
          description: When the event occurred
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_INPUT
                - UNAUTHORIZED
                - INSUFFICIENT_CREDITS
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Invalid or expired token'
...